# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.SwitchResDTO:
    properties:
      message:
        type: string
    type: object
  entity.Message:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/entity.StatusType'
      to:
        type: string
      updatedAt:
        type: string
    type: object
  entity.StatusType:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - Active
    - Sent
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/message/get-sent-messages:
    get:
      consumes:
      - application/json
      description: Retrieve a list of sent messages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Message'
            type: array
      summary: List messages
      tags:
      - messages
  /api/v1/message/switch-auto-messaging-mode/{active}:
    get:
      consumes:
      - application/json
      description: Start/Stop automatic message sending
      parameters:
      - description: Active Mode
        in: path
        name: active
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SwitchResDTO'
            type: array
      summary: Start/Stop Messaging
      tags:
      - messages
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"